refactor: rename TableOfContents to DomExtractedTableOfContents
