+++ src/components/core/navigation/DomExtractedTableOfContents.tsx
