+++ src/components/core/navigation/index.ts
@@ -1,2 +1,2 @@
 export * from "./ProductSelectors";
-export { default as TableOfContents } from "./TableOfContents";
+export { default as DomExtractedTableOfContents } from "./DomExtractedTableOfContents";
